refactor: make default metadata are linked to the environment

+++ gravitee-apim-rest-api/gravitee-apim-rest-api-management-v2/gravitee-apim-rest-api-management-v2-rest/src/main/java/io/gravitee/rest/api/management/v2/rest/resource/api/ApiMetadataResource.java
@@ -62,9 +62,7 @@ public class ApiMetadataResource extends AbstractResource {
                 MetadataResponse
                     .builder()
                     .data(MetadataMapper.INSTANCE.mapFromCore(paginationData))
-                    .pagination(
-                        PaginationInfo.computePaginationInfo((long) output.metadata().size(), paginationData.size(), paginationParam)
-                    )
+                    .pagination(PaginationInfo.computePaginationInfo(output.metadata().size(), paginationData.size(), paginationParam))
                     .build()
             )
             .build();
